// apps/frontend/src/types/Plant.ts
export type Plant = {
  name: string;
  id: number;
  scientificNames: string[];
  otherNames: string[];
  images: string[];
  cycle: string;
  watering: string;
  wateringGeneralBenchmark: string;
  sunlight: string;
  pruningMonths: string;
  pruningCount: string;
  flowers: boolean;
  floweringSeason: string;
  fruits: boolean;
  fruitingSeasons: string[];
  harvestSeasons: string[];
  harvestMethods: string;
  growthRate: string;
  maintenance: string;
  medicinal: boolean;
  edibleLeaf: boolean;
  poisonousToHumans: boolean;
  poisonousToPets: boolean;
  cuisine: boolean;
  indoor: boolean;
  careLevel: string;
  rareLevel: string;
  rare: boolean;
  description: string;
};